chore: guestToken 관련 파일들 위치 변경
